+++ benchmarks/historic/v1/src/astar.cpp
@@ -1,4 +1,4 @@
-#include "pathfind/astar.hpp"
+#include "astar.hpp"
 
 #include <cassert>
 
@@ -6,11 +6,6 @@ using namespace pathfind;
 
 Path Astar::calculatePath(const Graph& graph, node_id start_id,
                           node_id goal_id) {
-#ifdef BUILD_BENCH
-  nodes_expanded_ = 0;
-  edges_examined_ = 0;
-#endif
-
   // empty path to return on failure
   Path empty_path;
 
@@ -18,11 +13,14 @@ Path Astar::calculatePath(const Graph& graph, node_id start_id,
   open_list.clear();
   closed_list.reset();
   gScore.fill(INT32_MAX);
-  came_from_list.fill(start_id);
+  fScore.fill(INT32_MAX);
 
   gScore[start_id] = 0;
+
   uint32_t start_heuristic = euclidean_heuristic(graph, start_id, goal_id);
-  open_list.add(start_id, start_heuristic);
+  fScore[start_id] = start_heuristic;
+
+  open_list.add(start_id, fScore[start_id]);
 
   // expansion loop
   while (open_list.size > 0) {
@@ -32,36 +30,37 @@ Path Astar::calculatePath(const Graph& graph, node_id start_id,
       return reconstructPath(current.id, start_id);
     }
 
-    closed_list[current.id] = true;
+    const Node* node = graph.getNode(current.id);
+    if (!node) {
+      continue;  // todo: handle invalid node state (nullptr)
+    }
 
-#ifdef BUILD_BENCH
-    ++nodes_expanded_;
-#endif
-
-    for (const Edge& edge : graph.getNeighbors(current.id)) {
-#ifdef BUILD_BENCH
-      ++edges_examined_;
-#endif
+    // neighbor_id node expansion through linked list traversal
+    uint32_t edge_index = node->first_edge_index;
+    while (edge_index != 0) {
+      const Edge& edge = graph.edges[edge_index];
       node_id neighbor_id = edge.target;
 
       // skip if already in closed list
       if (closed_list[neighbor_id]) {
+        edge_index = edge.next_edge_index;
         continue;
       }
 
       // avoid int overflow
       if (gScore[current.id] == INT32_MAX) continue;
-
-      uint32_t tentative_g = gScore[current.id] + edge.cost;
+      int32_t tentative_g = gScore[current.id] + edge.cost;
 
       // new best path found
       if (tentative_g < gScore[neighbor_id]) {
         came_from_list[neighbor_id] = current.id;
         gScore[neighbor_id] = tentative_g;
-        int32_t tentative_f =
+        fScore[neighbor_id] =
             tentative_g + euclidean_heuristic(graph, neighbor_id, goal_id);
-        open_list.add(neighbor_id, tentative_f);
+        open_list.add(neighbor_id, fScore[neighbor_id]);
       }
+
+      edge_index = edge.next_edge_index;
     }
   }