+++ src/fixtures/static_mock_graphs.cpp
@@ -1,7 +1,5 @@
 #include "fixtures/static_mock_graphs.hpp"
 
-#include <memory>
-
 #include "pathfind/graph.hpp"
 
 namespace pathfind {
@@ -11,17 +9,29 @@ Graph const& createMockGraph(MockGraphType type) {
     Graph graph{};
 
     // NODE 0
-    graph.nodes[0] = {0, 0, 0, 2};
+    graph.nodes[0] = {
+        Coordinates{0, 0},
+        0,
+        2,
+    };
     graph.edges[0] = {1, 10};
     graph.edges[1] = {2, 20};
 
     // NODE 1
-    graph.nodes[1] = {10, 0, 2, 2};
+    graph.nodes[1] = {
+        Coordinates{10, 0},
+        2,
+        2,
+    };
     graph.edges[2] = {0, 10};
     graph.edges[3] = {2, 10};
 
     // NODE 2
-    graph.nodes[2] = {0, 20, 4, 2};
+    graph.nodes[2] = {
+        Coordinates{0, 20},
+        4,
+        2,
+    };
     graph.edges[4] = {0, 20};
     graph.edges[5] = {1, 20};
 
@@ -31,9 +41,18 @@ Graph const& createMockGraph(MockGraphType type) {
   static const Graph disconnected = [] {
     Graph graph{};
 
-    // nodes 0 and 1 have no outgoing edges
-    graph.nodes[0] = {0, 0, 0, 0};
-    graph.nodes[1] = {10, 0, 0, 0};
+    // Nodes 0 and 1 have no outgoing edges.
+    graph.nodes[0] = {
+        Coordinates{0, 0},
+        0,
+        0,
+    };
+
+    graph.nodes[1] = {
+        Coordinates{10, 0},
+        0,
+        0,
+    };
 
     return graph;
   }();