+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Expanded A* benchmarks
 * Memory usage report accessable through `make memory`
 * `generate_config_header.py` generates pre-compiled headers for `config.json`
+* Add historical benchmarking
+* Create `Graph` interface
 
 ## Release date: 2026-05-26 Pico Router 0.1.0
 * Docker devcontainer